Rescue Uphold invalid_grant error to clear uphold_code
Handles the special Faraday::ClientError representing the invalid_grant case where the code is invalid (timeout or already used).
Clears the code when detected.

On other Uphold connection errors the code is preserved so the code exchange can be tried again.
Note: relying on Uphold rejecting in invalid code in the future to clear the code.

(cherry picked from commit 29f2dc5)