Handles protocols when registering a new publisher
When the publisher domain is being normalized any provided protocol is
stripped and replaced with `http` before being sent to
`ledger /v2/publisher/identity`.

Handles the case where a domain is malformed, such as including spaces,
before sending to be normalized. This prevents sending domains to the
ledger which are not valid domains. It also adds a further degree of
consistency between the online and offline modes.

Adds testing for the above and basic mock testing for online mode.

Fixes #53