Allow DNS verify form to work with JS disabled
The prior approach used a hidden form triggered by a `click` event on an
independent button. This did not work with JS disabled.

This approach converts the independent button to be a submit button for
a standard (non-hidden) form that will submit regardless.

The display of the interstitial should still happen with JS enabled.

Further exploration of an XHR-based alternative that uses polling is 
recommended.

[Resolves #43]