Dynamically generate publisher statements from dashboard
Provides a `select` control to choose a period, and a `Generate` link
that invokes an Ajax call to the new `generate_statement` endpoint.

The report URL that’s returned will be polled periodically to see if
it exists (200 is returned). Polling is done every 3s with a 2s 
backoff period. Four attempts are made.

If the report has been generated, a link will be presented to view it.

If the report is not generated during the polling period, the user is
told that a link will be emailed when the report is ready.