Show Uphold status + connect button on dashboard
For now the Uphold connection button is only shown when `uphold_status`
is either `unconnected` or `code_acquired`. The only way to get into 
the `code_acquired` state is for the code exchange to have failed, 
so we should immediately allow for user’s to connect a new account
(since the code expires in 5 minutes).