Simplify Publisher#prepare_uphold_state_token
An uphold_state_token is required in order to generate the URL for 
Uphold authorization. 

Publisher#prepare_uphold_state_token now generates a fresh 
uphold_state_token _only if one is not set already_, and it no longer
changes any other attributes on the publisher.

As a result, `Publisher#prepare_uphold_state_token` can now be called
idempotently from the `uphold_authorization_endpoint` helper, just to
guarantee that an `uphold_state_token` has been set.

This simplifies the logic around setting state tokens, which allows
Uphold tokens to be reset in case of problems and in case a user wants
to change their Uphold account.

Note that whenever an Uphold code is obtained, the state token is 
still cleared, which guarantees that a fresh state token will be used
to obtain the next code (if necessary).