Introduce new “dynamic ellipsis” for showing progress
When generating a report or showing uphold statuses that are 
“processing”, it’s useful to show progress more subtly than with a
spinner that overlays the page. The rationale is that these processes
may take a while and should not dominate the UI while they are 
occurring in the background.

This technique simply increments the number of dots display after a
status message (like “generating…”), updating every hundred ms or so
and then gradually resetting back to 0 dots before starting again.

For example:
Generating
Generating.
Generating..
Generating...
Generating....
Generating
Generating.
Generating..
Generating...
Generating....