Ensure publisher's status is set correctly on the dashboard (#156)
* Refactor calculation & display of Publisher status

The helpers used to determine a publisher's status, status description,
and next step path have been refactored to eliminate redundancy and 
unnecessary cases.

The publisher's status icon on the dashboard is now controlled by an
element class rather than an inline image. This will make it easier to
change statuses dynamically.

* Replace publishers/uphold_status with publishers/status

This more general route now includes both the publishers general status
and description, as well as the uphold status and description.

This route is invoked from the Dashboard so that both the uphold status
and general status (and icon) can be updated together.