List statements on dashboard
The statements are shown in chronological order, newest at the top.

When `Generate` is clicked, a new row is added to the top, and 
`Generating…` is displayed in the `Download` column. Once successful, 
this becomes a `Download` link like the others.

Only one report can be generated at a time. After the report has been 
generated, the `Generate` link reappears.

If the report is not generated after > 1 minute of polling, `Delayed`
is displayed.

Closes #208