Ask publishers for listing permission immediately after verification (#258)
* Remove show_verification_status checkbox from contact info form

This was deemed too early in the process to ask for this permission.

* Show modal immediately to confirm verification status selection

Immediately after a publisher has been verified, show a dialog that asks
them to confirm whether they want to appear as a verified brave 
publisher.

This dialog is only shown when `show_verification_status` is `nil`.

* Allow `publishers.show_verification_status` to be null

No longer set this field to be `true` by default. The presence of `null`
indicates that a choice still needs to be made.

* Highlight noscript warning more distinctly

* Ensure that modal and overlay are noscript compatible

The verification modal won’t be displayed at all with JS disabled.

This requires usage of `display: none` by default.