Allow publisher domain to be set sync or async
If JS is enabled, the contact info form will be submitted via xhr and
the publisher's domain will be set asynchronously. The domain status
will be polled repeatedly and then the page will be advanced when the
domain has been set.

If JS is disabled, the contact info form will be submitted via a
standard http PATCH and the publisher's domain will be set
synchronously.