release 2017-12-06 (#361)
* Logging: staging logs to INFO level, add LOG_API_REQUESTS

* use deliver_later instead of deliver_later!

* Provide link to sign up within flash message if publisher attempts to sign with unregistered domain

* Separate en.publishers.new_auth_token_publisher_not_found into two parts, the second of which will be the link

* Modify find_publisher to pass both parts as the error message to the publisher controller instead of the removed translation

* Modify publisher controller to flash both messages, adding a link to the homepage for the second

Refs #325

* Use partial to render a new :login_link flash message instead

* No longer use error message passed to the publishers controller, instead use new :login_link flash message

* Change `find_publisher` method to not pass error message, only return false if no publisher found

* Create new "login_link" partial to be rendered instead

* Add a new flash message :login_link that renders "login_link" partial

* Add client side validations for publishers phone numbers

* Add non-strict validation that forces phone numbers to only include numbers, dashes, parentheses, periods, and plus (+)

* Add title field within html tag to give user more information if they enter an invalid phone number

* Add the title message to en.yml for internationalization

Refs #324 and #329

* Run the development server as SSL.

As we add support for U2F-based 2FA and adopt other cutting-edge web
features, we need to be working on an HTTPS domain. This is because some
APIs (like U2F) are only available to an SSL host.

Please create an entry in your `/etc/hosts` file for `localhost.ssl`
pointing at `127.0.0.1`. The following command should be sufficient:

```
echo "echo '127.0.0.1 localhost.ssl' >> /etc/hosts" | sudo -s
```

The built-in development Puma server expects two files to exist in the
`ssl/` directory:

```
ssl/server.key
ssl/server.crt
```

These should be a key and certificate for the domain `localhost.ssl`.
You may already have certificates for this domain, and re-using those is
encouraged.

However if you need to generate a key and certificate, you can use the
following commands on OSX:

```
bundle exec rake ssl:generate
```

Will generate a key, signing request, and certificate.

```
bundle exec rake ssl:install
```

Will add the certificate to your OSX keychain.

Once you've run these commands you can simply run the Foreman server
like normal and visit https://localhost.ssl:3001/. You may still need to
approve an exception to allow the certificate the first time you use it
in a browser.

References: https://github.com/brave-intl/publishers/pull/323

* Correct the translation to be flashed when recaptcha verification fails

Refs #333

* Use localhost:3000, no OSX-specific cmds

* Tweak font sizes on mobile

Refs: https://github.com/brave-intl/publishers/issues/199

* Upgrade ruby to 2.3.5

Fix #338

* Use normalized domain for finding publishers during login

Fixes #336

* Extract spinner from dashboard into separate module

* Extract fetch utils from dashboard

* Remove unused domain normalization code

* OfflineNormalizationError is not used
* PublishersController#normalize_brave_publisher_id is never invoked

* Introduce flash.js

Simple control of the flash element from JS that’s compatible with
server-side rendering of the flash.

* Add domain_normalization_status to publishers table

This allows domain normalization to proceed asynchronously and for the
results to be persisted and polled.

* Validate publisher domains based on domain_normalization_status

Instead of normalizing the domain and inspecting the host as a
synchronous part of validating a Publisher, rely on outside services
to do these tasks (preferably asynchronously) and then set
domain_normalization_status in case of a problem.

* Introduce PublisherDomainSetter service

This service takes a publisher and a domain, normalizes the domain,
inspects the host, and then sets `publisher.brave_publisher_id` on
success.

This processing was formerly performed in the Publisher model.
Separating this out into a service allows it to happen synchronously
or asynchronously (via a job).

* Introduce SetPublisherDomainJob

Invokes PublisherDomainSetter service to normalize and inspect a domain.

* Introduce PublishersController#domain_status

Endpoint is used to poll the status of setting the domain on a
publisher. Normalization and inspection errors will be reflected as
a successful (200) response that contains an `error` element.

* Allow publisher domain to be set sync or async

If JS is enabled, the contact info form will be submitted via xhr and
the publisher's domain will be set asynchronously. The domain status
will be polled repeatedly and then the page will be advanced when the
domain has been set.

If JS is disabled, the contact info form will be submitted via a
standard http PATCH and the publisher's domain will be set
synchronously.

* Clarify @ symbol for DNS site verifcation

* Remove 'Name' form with @ symbol that may confuse users

* Replace with instructions on adding to root domain, clarify this is often represented by @

* Downcase the 'P' in 'Copy and Paste'

* Add translation for 'Copy to Clipboard' and downcase the 'C'

Refs #135

* Allow developers to force a Captcha using query string param (#344)

* Allow developers to force a Captcha using query string param

* Add additional case for turing on captcha, when params[:captcha] == true

* Add hidden field tag to captcha protected forms that passes the captcha param into the POST

Refs #339

* Alter StaticController so captcha appears upon redirect

* Add @should_throttle instance variable to StaticController#create to enable recaptcha tags

* Copy over `should_throttle_create?` from PublishersController to StaticController

* Assign @should_throttle in StaticController using `should_throttle_create?` or `params[:captcha]`

* Remove publishers.verification_failed translation and references to it, as it would appear before the captcha is filled out.

* Add check for params[:captcha] within views before creating the hidden field

* When checking @should_throttle, check only for existance of params[:captcha] instead of checking for "true"

Refs #339

* Style change

* Use newer colon syntax for captcha param on redirect to root_path for PublishersController#create

* Simplify and improve statement generation (#347)

* Notify admins when publisher emails change (#350)

* Add confirm_email_change_internal message to the publisher mailer

* Add unit tests for exisitng confirm_email_change message in publisher_mailer_tests.rb

* Add test for the entire changing email flow in publisher_controller_tests.rb

Refs #116

* Refactor validations related to brave_publisher_id (#351)

* Ensure resource strings for brave_publisher_id errors are consistently formatted

* Clarify publisher domain validations

* The PublisherDomainSetter can not allow duplicate domains, and must
  set brave_publisher_id_error_code so that this scenario will be 
  properly displayed to the user.

* A publisher record CAN NOT be valid when it has a brave_publisher_id
  and either a brave_publisher_id_error_code or a 
  brave_publisher_id_unnormalized.
  
* Changing brave_publisher_id_unnormalized should clear 
  brave_publisher_id_error_code.

* The SetPublisherDomainJob must clear brave_publisher_id_unnormalized
  so that publishers with a brave_publisher_id_error_code
  will be valid when saved.

* gems: update redis-rails and redis-store (#354)

fix #353

* Add locale label for brave_publisher_id_unnormalized (#357)

* Add escape hatch link for misidentified WordPress sites (#359)

* Update wording and format for Wordpress escape hatch (#360)