Add agreed_to_tos to publishers
The agreed_to_tos timestamp marks the first time that a publisher has completed signup and thus explicitly agreed to the site’s TOS.

Publisher verification, and thus access to the dashboard, will be blocked on the presence of agreed_to_tos. Thus, all returning publishers will be presented with the complete signup page.