Show thumbnail images for channels
If an image is not provided for a channel (or channel type), then a default image will be used.

We may wish to consider reformatting the default channel image to work displayed as a circle, so that we don’t have a mix of square + circular thumbnails. Even though the default image isn’t obviously squared off, keeping things consistent will be easier.