Enable channels to be removed (#501)
* Enable channels to be removed

A `remove` link per channel on the dashboard deletes the channel with an xhr and, if successful, animates the channel row off the page. A confirm dialog is used to ensure that the publisher really wants to delete the channel.

The ChannelsController is updated to deal with xhr json requests and no longer deals with redirects or flash updates.

closes #385, closes #413

* Introduce PublisherChannelDeleter service

* Introduce DeletePublisherChannelJob

Performs PublisherChannelDeleter service.

* Deleting a channel triggers a DeletePublisherChannelJob

DeletePublisherChannelJob syncs the job deletion with eyeshade.