Show modal when channel has already been claimed
Restores the modal dialog that’s shown when a YouTube channel has already been registered. This dialog has been enhanced to work with web site channels as well. Both versions also have a `noscript` equivalent.

For YouTube channels which have already been registered, the person who registered the channel and the registration date are shown. This has been seen as acceptable since the user will not see this message without already authenticating themselves as a manager for the channel.

For web sites, the registration information is omitted because no authorization step is performed prior to showing the “channel taken” dialog.