Push default_currency updates to eyeshade
When the default_currency is successfully updated, an UploadDefaultCurrencyJob will be triggered that will upload the change to eyeshade.

The UX has been tweaked slightly so that, for users with no default currency, a select option `— Select currency —` will appear as selected by default. This allows any other selection to trigger a change (since there’s no default chosen as was previously the case).