Clarify Uphold reauthorization status
When eyeshade responds that Uphold authorization / reauthorization is needed, the `uphold_verified` status should not be revoked immediately, because this status can be resolved either by a user or by Uphold without involving Brave. Instead `:reauthorization_needed` should be viewed as a a substate of `:verified`.

This approach also fixes the case in which a user has explicitly disconnected their Uphold account. They will always be shown a status of “Not connected” and never be shown “Connection problems”.

Fixes #841