Better handle scenarios in which a publisher statement is empty
If a publisher’s statement is returned as an empty string, then we don’t want to notify the publisher that it is ready.

We need to check that `contents.present?` instead of just the truthiness of `contents`. In this way, the same logic is applied in both PublisherStatementSyncer and SyncPublisherStatementJob for evaluating the presence of a statement’s contents.

This fixes an error that resulted because eyeshade was returning statement URLs that required a redirect to fetch.