Introduce PublishersController#confirm_default_currency
Allow xhr patches to update a publisher’s default_currency.

If the user is able to create Uphold cards, create a card immediately and return a response indicating that the page should be refreshed.

If the user needs Uphold authorization, respond with a redirect URL and status message for the client.

Once the user returns `home`, create Uphold cards in the background if necessary.