Introduce new `incomplete` Uphold status
The `incomplete` status represents the case in which a publisher has not finished registering their account at Uphold. They may need to verify their email or configure 2FA. An incomplete registration will be unable to create currency cards, so we need to wait for their registration to be complete before displaying the currency selection modal.

An `incomplete` status can be detected from the wallet returned from eyeshade. Its `authorized` field will be `false`.