WIP: Introduce dashboard charts
This introduces the dashboard charts described in https://github.com/brave-intl/publishers/issues/542

* A donut chart of contributions
* A stacked bar chart of monthly deposits

These charts are written to be customizable, but still require data in a certain schema (i.e. format). A common array of channels and corresponding colors can be passed, as well as a number of options related to the display.

The charts themselves are rendered with D3.

An example data set is used to explore usage of the charts on the dashboard. To move beyond this WIP, we’ll need to invoke rendering based on live data coming from eyeshade.

In order to enable display of these charts on the dashboard, visit `/publishers/home?charts=true` with the temporary query parameter.