Add 'active' column to PromoRegistration
Admins now have have the ability to pause reporting
for a given PromoRegistration.  If a code is not active
and someone downloads using the code, it will be not
be counted.  This bit of state must be tracked on the
referral server but it will also be stored on publishers.

Also check in db/schema.rb