Add rake task to backfill selected_provider_provider_*
The task looks for all publishers that don't have a selected
wallet provider set on the publishers record, then for each of those
checks first if there is a gemini connection, second if there is a
paypal connection and third checks if there is an uphold_connection.  If
there are it sets the selected_wallet_provider_type and
selected_wallet_provider_id fields in the publishers record.

If there isn't any of gemini, paypal, or uphold connections, nothing is
set.