Further consolidates Bitflyer Oauth refresh
I've now eliminated most of the code in the old auth.rb which
wasn't being used. Since we're just using that for HTTP calls
and the business logic is elsewhere, I've renamed the file http
and moved the routes information there. This let me delete the
bitflyer.rb main module. Now we're down to 2 files to look at
to understand Oauth refresh: the service file and the http file.