Throw an exception if redis down in development
When testing a recent PR in dev, I was not connected to
redis. Instead of give me a warning, the app switched to
null store which caused me to waste a bunch of Albert's
time asking why a PR didn't work.

With this change, if we can't connect to redis, we throw
an exception. Also, in general, if there are any issues with
redis this will also throw an exception which is desirable in
development.