Upgrade packages to address is-svg vulnerability
We have a couple of dependabot vulnerabilities listed, so
I updated the rails-webpacker package which led to is-svg
being eliminated as a dependency. I'm still waiting on the
ssri fix https://github.com/rails/webpacker/issues/2969
and perhaps that will not come until the rails-webpacker
gem get a 6.0 release, which should be sometime soon.