Update Uphold Payout Test to cover BitFlyer
Paypal is going away, and we can remove that code next month
after the next payout. We also want to clean up the database
to ensure that every creator has a selected_wallet so that we
can simplify the service.rb skip_publisher? method, as well as
any other code would benefit from the data cleanup.