Enable Bitflyer potential payments (#3205)
* Enable Bitflyer potential payments

We can now generate potential payments fot Bitflyer creators.
Note that I tried to clean up a little of the calling code,
as it appears we no longer need should_send_notifications.
Also the class is now easily testable without using web mocks
since we can pass in the refresher.

* Return to keyword args

* Restore should_send_notifications, remove later

* Test update

* Removing unneeded wallet checks

* Update tests

* Restore should_send_notifications until a later PR refactors

* Update Uphold Payout Test to cover BitFlyer

Paypal is going away, and we can remove that code next month
after the next payout. We also want to clean up the database
to ensure that every creator has a selected_wallet so that we
can simplify the service.rb skip_publisher? method, as well as
any other code would benefit from the data cleanup.