Fix setting the JA locale in the URL
Previously we could set locale=JA in the URL and see pages in
Japanese, but broke that feature recently. This restores it and adds
appropriate testing. Note that you can use a browser sending an
English accept header and set this locale and have it work. If you
are using a browser where Japanese is the preferred language,
you will automatically see Japanese pages. You can not yet manually
set the locale to English if we detect you are using an JA broswer,
but we could add it later if pressed.