Fix setting the JA locale in the URL (#3224)
* Fix setting the JA locale in the URL

Previously we could set locale=JA in the URL and see pages in
Japanese, but broke that feature recently. This restores it and adds
appropriate testing. Note that you can use a browser sending an
English accept header and set this locale and have it work. If you
are using a browser where Japanese is the preferred language,
you will automatically see Japanese pages. You can not yet manually
set the locale to English if we detect you are using an JA broswer,
but we could add it later if pressed.

* Remove launchy due to need for security review

* Run JP Web tests non-headless

* Use FF for headless Japanese feature test

Chrome currently has a bug in linux where it does not respect
the preferred language in headless mode. The easiest workaround
for the time being is to use FF.
See https://bugs.chromium.org/p/chromium/issues/detail?id=1010288