Enable CSP policy in reporting mode (#3270)
* Enable CSP policy in reporting mode

We want to ensure a good CSP policy. This was initially driven by
the need to look at spectre mitigations (#3124), but as part of that
involves a good CSP policy, i.e.
https://security.googleblog.com/2020/07/towards-native-security-defenses-for.html

we need to get a good CSP policy in place to begin with.

For now, we'll just log the violations. We can quickly clean them up
and keep an eye on any new violations that come in over time, eventually
phasing out this code.

* Add test for Csp Violations