Rotate attr_encrypted (#3320)
* Draft rotate attr_encrypted

* Added rollback if new value doesn't match old

* Rekeying now works

In order to do so, in the secrets.yml file, set the current
attr_encrypted_key old field to attr_encrypted_key_old and then
put the new key in attr_encrypted_key. Then run `rake rekey`.

Note that this does this one record at a time and we do not
yet handle reading the old value from fields while rekeying,
as rekeying could take a while. We need to add that support
in and then /perhaps/ we can refactor this code to work in a
bulk update manner, which would be faster.

Note that now we pass in the key: to encryption_key for all the
models since some of them use the old slice method on the keys
and some use the new pack. Getting them all on the same method
is out of the scope of this PR. Note also that I've made all
of the encryption_key methods class methods for consistency
and so that we don't have to repeat code unnecessarily,
espcially now that we pass the key in.

* Add support for reading old values from the DB while rotating

* Add Bulk Update support, don't rekey UserAuthenticationToken

* Don't monkey patch in testing

* Remove payout_reports content_rekeyed

* Rekey without checking DB; skip if done before

* Remove migration

* No need to set rekeyed_at time

* Add ordering for rekey task

* Get old key from env