Wasmer image (#3335)
* Add WASM image resize support

This lets us run rust code to resize images in a WASM environment.
We can now also remove minimagick.

The output image is a jpg. We may want to investigate converting
to webp in the future so that all images will get transformed
into a new format.

The wasm binary comes from https://github.com/brave-intl/wasm-thumbnail

* Fix tests

* Remove monkey patch of Active Storage

Since Rails 6.1 we can specify a service-per-attachement, we
do that and get rid of the monkey patch we used before to
achieve the same.

Note we default to the case_notes bucket and only specify the
service for the banner images.

* Fix tests, deallocate memory

* Keep image padding, don't recompile wasm more than once

* Add rust to Dockerfile

* Use newest version of wasm-thumbnail in Gem with register_panic

* Rely on gem

* Update gem

* Move resize method to gem

* Rename buckets

* Reference tag for thumbnail gem

* Allow larger image uploads and updage gem due to bug

In the gem, the panic handler wasn't registered properly.
I also noticed that if we attempt to resize a larger image,
the gem panics because we didn't give it enough space. I've
added more space/memory to allow for this.

* Use target file size

* Add appropriate error message for large images

* Remove wasm binary

* Add adjustable quality for image resizing

* Re-enable size test

* Use right variable

* Fix test for max image size error