Have docker cache wasmer before installing
Wasmer has a lot of trouble properly installing
in our Docker environment when we are using deployment
mode. If not deploying, the asset compile can't find
it. So we use deployment mode, but have to cache the wasmer
gem before we do that or we get a git error due, I suspect
due to the gem not being formed properly, but wasn't able to
ascertain the exact reason.