Initial stab at simplifying potential payments
We want to bulk insert these for performance reasons. We no longer send
out emails and with this, we will no longer sync connection information
for each of the wallets. That is done when a user logs in.

This lets us delete and simplify the code. I'm sure there are things
that can still be done, and might be -- but this is a first stab.

Things to be done for this PR before merge:
- tests pass
- Make sure all records loaded up front, so no n+1

Things that still might be done, but maybe for v2:
1) Eliminate the create_message calls and just return the objects
   for bulk insertion
2) Make uphold_service and gemini_service style like bitflyer_service