Fix asset compilation
This was failing previously because we were attempting to
compile assets in the dev environment when not all those
dependencies are available since we install gems only in
the prod environment. To run in the prod environment
without variables set, we needed to get rid of an old
check that only ever checked a single variable. We also
needed to set the secret key, but that's all that was
needed.