Fix Potential Payments ordering bug (#3441)
* Fix Potential Payments ordering bug

When generating the payout report, the publisher statuses
were not being properly accounted for after switching to the new
much speedier, eager loaded payout report generation. When eager loading,
Rails was dropping the default ordering on the publisher statuses
table and instead just ordering by ID. I played around with added a new
scope that would only let us pull in the latest status update. The scope
looked like this:

  # Eager loading these will NOT keep the ordering, so make it explicit here
  has_one :latest_status_update, class_name: "PublisherStatusUpdate", primary_key: :latest_publisher_status_id, foreign_key: :publisher_id
  scope :with_latest_status, lambda {
    select(
      "publishers.*,
      (
        SELECT publisher_id as latest_publisher_status_id
        FROM publisher_status_updates
        WHERE publisher_id = publishers.id
        ORDER BY created_at DESC
        LIMIT 1
      )"
    )
  }

But in the end, Rails kept wanting to use this scope in all subqueries
for processing the payouts. It looked like it wasn't going to work unless
I wrote a bunch of manual SQL, so instead I opted to separate the
payment generation into 2 steps.

1) Get all the publishers who need a payout, with all the crazy scoping
that entails.
2) Eager load those publishers.

By removing the crazy scoping and doing a simple query to eager load,
Rails keeps the default ordering. In particular, the parts of the query
that overrode the status ordering was

`with_verified_channel.not_in_top_referrer_program.valid_payable_uphold_creators`

Which generates all kinds of subqueries that make it tricky to try to
enforce the proper ordering of the publisher statuses.

* Disable rubobop for 2 lines

* If there is no authorization time, refresh

With some of the tokens that Uphold gave us, there was no
expiration time set. We should refresh if that is the case.