Add Wallet Syncer which Updates on Login
We don't want to update on home page view, and we want to cover
all wallet providers. Currently we won't sync Bitflyer since I'm not
entirely sure of the stability of that part of the codebase and this is
mainly addressing Uphold, for the moment. Will need to come back and
turn that on for Bitflyer, though it's not as big of a priority since
at connection, the user is already KYCd.