We were over reporting the nightly stats email
The promo system doesn't set the finalized=true until 30 days have passed,
so this nightly report will only report up until a month ago. That is,
where a
retrieval = download from site
download = actual download and then launch

The Referral download table is a view on the download table, that is,
first launches. That table also tracks usage pings so a column gets updated
daily in it. That column is the finalized_ts, which is very unfortunate
and in fact, much of the naming here is.

So we get a finalized_ts that gets set daily, but the finalized column
itself doesn't get set until a month later.

So if you are looking at this report on Dec 6, the report's latest date
will be Nov 5, i.e. when a particular download was created. It's finalized
column was updated today, and it's finalized_ts column has today's date
in it, but all that matters is the `ymd`, that is, the created date, and
the finalized boolean. This finalized boolean is also what the promo UI
keys off of, so this brings the report in line with the UI.