Resolved 504s during pCDN upload
The call to find the SiteBannerLookup

https://github.com/brave-intl/publishers/blob/staging/app/jobs/cache/browser_channels/responses_for_prefix.rb#L23

`@site_banner_lookups = SiteBannerLookup.where("sha2_base16 LIKE ?", prefix + "%")`

was doing a table scan and not using the index. Through experimentation,
I found out that we aren't using the current indexes because the collation
is off. This corrects that. I ran a full upload to the pcdn and verified
that the 504s are gone. This should greatly improve responsiveness both
for the admin panel and stats hitting our API.