Builder Base Class changes
We introduce a union type which can be distinguished by pattern matching

The reason for this is we want Sorbet to help us as much as possible. If
we have an error, we don't want to accidentally put those in a success
callback. With the fields of the success/error structs being different,
sorbet can now help prevent sensitive information escaping.

An example of usage:

  case res = ChildService.build.call
  when BFailure
    puts "I can access errors!"
    problem("1")
  when BSuccess
    puts "I can access result!"
    pass(res.result.add(2))
  else
    T.absurd(res)
  end