This adds the ability to add a security key using webauthn (#3541)
* This adds the ability to add a security key using webauthn

The old U2F protocol is being phased out. The new Webauthn protocol is
more flexible and nicer to work with. Since most of the registrations are
in U2F, we just keep them there but mark the format. All new registrations
will be done using the Webauthn format. Note we must run the u2f rake job
in prod after deploy to mark them and standardize the base64 encoding of
the already-stored keys.

Tested locally with both a Yubikey that was registered using U2F and a
separate one under Webauthn. With both registered I could log in using
both.

* Extract Webauthn registration into service

* Add tests for Webauthn registration

* Fix tests / types

* Add CREATORS_HOST

* Env change

* CREATORS_HOST fix

* Type fix

* Authorization to new format