feat: Eyeshade::Balances
This is me taking everything I learned about Eyeshade and sorbet and
generating a new "Wallet" class.  I was pulled deeply down the rabbit
hole for this bat negative value issue and in an effort to show
something for it I've gone ahead and pulled this out of the large number
of things I uncovered as I made my way through the service stack.