feat: Oauth2::AuthenticationCodeBase (#3623)
Implements an abstract base class with a standard API for implementing
the oauth2 authorization code flow for a model that is a token
container.  Can be used for any wallet provider and is implementation
agnostic. The Oauth2::AuthenticationCode client can be extended to be
used for all oauth2 authentication code flows and replace all the
existing implementations of code requests, access token retrievals, and
refresh token flows.