refactor: Remove notification
Initially I believed this may be useful, but I doubt we would actually
use the notification for this particular job and I felt that it was
making things more and more difficult to reason about.  I think keeping
it abstracted from the ASYNC task is good because it helps with sorbet
typing and unit testing.  Sidekiq behavior doesn't play well with
sorbet.