Use WHATWG URL API rather than deprecated url.parse
url.parse uses a lenient, non-standard algorithm for parsing URL strings.
It is prone to security issues such as host name spoofing and incorrect handling of usernames and passwords

This change maintains the previous behaviour of stripping the querystring
and hash from the URL before redirecting (just using the path).