First pass at getting new UI testing down
For now we have to run next spearately to run the
tests, eventually we want capybara to launch next
and run the tests.

At least locally, this gets

bin/rails test test/features/publishers_home_test.rb:28

working. The shadowdom is tricky to work with, so
we might want to revisit that in the future.